Delete CommentsConcern
